File rename | Renames of the available natives | Props are now deleted on scenario stop
